Renamed Card -> GameCard. #110
